requirements.txt
arxiv
schedule
PyYAML

# Testing
pytest
pytest-mock
